// src/app/api/v2/model/externalUserChangeResponse.ts
/**
 * KITOS API V2
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { APIExternalUserChangeType } from './externalUserChangeType';
import { APIExternalUserChangeStatus } from './externalUserChangeStatus';


export interface APIExternalUserChangeResponse { 
    uuid?: string;
    externalMessageId?: string | null;
    externalUserUuid?: string;
    userUuid?: string | null;
    userName?: string | null;
    email?: string | null;
    changeType?: APIExternalUserChangeType;
    status?: APIExternalUserChangeStatus;
    occurredAt?: string | null;
    receivedAt?: string;
    resolvedAt?: string | null;
    resolvedByUserUuid?: string | null;
}