// src/app/api/v2/model/externalUserChangeStatus.ts
/**
 * KITOS API V2
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type APIExternalUserChangeStatus = 'Pending' | 'Applied' | 'Dismissed';

export const APIExternalUserChangeStatus = {
    Pending: 'Pending' as APIExternalUserChangeStatus,
    Applied: 'Applied' as APIExternalUserChangeStatus,
    Dismissed: 'Dismissed' as APIExternalUserChangeStatus
};